refactor: update imports and improve consistency across components
- Refactored import paths to align with the new sfcc structure in multiple components.
- Standardized string formatting in the search page for improved readability.
- Adjusted schemas and utility imports for better organization.
