+++ app/search/page.tsx
@@ -1,11 +1,11 @@
-import Grid from 'components/grid';
-import ProductGridItems from 'components/layout/product-grid-items';
-import { getProducts } from 'lib/sfcc';
-import { defaultSort, sorting } from 'lib/sfcc/constants';
+import Grid from "components/grid";
+import ProductGridItems from "components/layout/product-grid-items";
+import { defaultSort, sorting } from "lib/constants";
+import { getProducts } from "lib/sfcc";
 
 export const metadata = {
-  title: 'Search',
-  description: 'Search for products in the store.'
+  title: "Search",
+  description: "Search for products in the store.",
 };
 
 export default async function SearchPage(props: {
@@ -16,14 +16,14 @@ export default async function SearchPage(props: {
   const { sortKey, reverse } = sorting.find((item) => item.slug === sort) || defaultSort;
 
   const products = await getProducts({ sortKey, reverse, query: searchValue });
-  const resultsText = products.length > 1 ? 'results' : 'result';
+  const resultsText = products.length > 1 ? "results" : "result";
 
   return (
     <>
       {searchValue ? (
         <p className="mb-4">
           {products.length === 0
-            ? 'There are no products that match '
+            ? "There are no products that match "
             : `Showing ${products.length} ${resultsText} for `}
           <span className="font-bold">&quot;{searchValue}&quot;</span>
         </p>
